Added VSCode helper items, very basic barebone code started
The idea right now is to have the tool search for the item, find the one
that best matches and add that item to the cart.

The JSON has not been flushed out yet but it can potentially have fields
that make some functionality redundant. However the idea that you can
filter my a price range or a rating range or both is a good starting
place.